Move from a placeholder (fake) API key to an external API file
Previously, the `app/build.gradle` file was checked into version
control with a placeholder string for the API key. For ease of version
control, the key location has now been moved to a single-purpose
`app/api.gradle` file that isn’t checked in. As per #1, instructions on
adding this file should documented.